Move DecodeStream and LiftReader to io.finch package
