+++ circe/src/main/scala/io/finch/circe/AccumulatingDecoders.scala
@@ -7,9 +7,8 @@ import cats.data.Validated
 import io.circe._
 import io.circe.iteratee._
 import io.circe.jawn.{decodeAccumulating, decodeByteBufferAccumulating}
-import io.finch.{Application, Decode}
+import io.finch.{Application, Decode, DecodeStream}
 import io.finch.internal.HttpContent
-import io.finch.streaming.DecodeStream
 import io.iteratee.{Enumeratee, Enumerator}
 
 trait AccumulatingDecoders {
