+++ circe/src/main/scala/io/finch/circe/Decoders.scala
@@ -7,9 +7,8 @@ import io.circe.Decoder
 import io.circe.fs2
 import io.circe.iteratee
 import io.circe.jawn._
-import io.finch.{Application, Decode}
+import io.finch.{Application, Decode, DecodeStream}
 import io.finch.internal.HttpContent
-import io.finch.streaming.DecodeStream
 import io.iteratee.Enumerator
 import java.nio.charset.StandardCharsets
