+++ core/src/main/scala/io/finch/internal/ToEffect.scala
@@ -1,13 +1,12 @@
 package io.finch.internal
 
 import cats.effect.Effect
+import cats.~>
 import com.twitter.util.{Future => TwitterFuture, Return, Throw}
 import scala.concurrent.{Future => ScalaFuture}
 import scala.util.{Failure, Success}
 
-trait ToEffect[F[_], G[_]] {
-  def apply[A](f: F[A]): G[A]
-}
+trait ToEffect[A[_], B[_]] extends ~>[A, B]
 
 object ToEffect {
 
@@ -16,14 +15,13 @@ object ToEffect {
   }
 
   implicit def twFutureToEffect[E[_]: Effect]: ToEffect[TwitterFuture, E] = new ToEffect[TwitterFuture, E] {
-    def apply[A](a: TwitterFuture[A]): E[A] = {
+    def apply[A](a: TwitterFuture[A]): E[A] =
       Effect[E].async { cb =>
         a.respond {
           case Return(r) => cb(Right(r))
           case Throw(t) => cb(Left(t))
         }
       }
-  }
   }
 
   implicit def scFutureToIO[E[_]: Effect]: ToEffect[ScalaFuture, E] = new ToEffect[ScalaFuture, E] {
